REFACTOR: split post_common_3d.py application (#5955)
