+++ doc/changelog.d/5955.miscellaneous.md
@@ -1,0 +1,1 @@
+split post_common_3d.py application
